Fix rails test command in CI
Since it was running the docker-compose exec command directly it never
had a chance to execute the webpacker:compile command that automatically
runs when you run rails test through the rails function that exists in the
run script.

This caused a `Webpacker can't find application.css` error because the
manifest was empty.